// src/interfaces/cards/Cards.ts
import { Address } from "./Address";
import { Image } from "./Image";

export interface Card {
  _id?: string;
  title: string;
  subtitle: string;
  description: string;
  price: number;
  rooms: number;
  category: string;
  phone: string;
  email: string;
  web: string;
  image: Image;
  address: Address;
  bizNumber?: number;
  likes?: string[];
  user_id?: string;
}